WIP extract classes
