+++ ssh-control-mcp/src/ssh/errors.ts
@@ -1,0 +1,10 @@
+/**
+ * Custom error class for SSH operations
+ */
+export class SSHError extends Error {
+  constructor(message: string, cause?: Error) {
+    super(message);
+    this.name = 'SSHError';
+    this.cause = cause;
+  }
+}
