+++ ssh-control-mcp/src/ssh/index.ts
@@ -1,0 +1,17 @@
+/**
+ * SSH module - Manages SSH connections and persistent sessions
+ */
+
+export { PersistentSession } from './session.js';
+export { SSHConnectionManager } from './manager.js';
+export { ConnectionPool } from './connection-pool.js';
+export { SSHError } from './errors.js';
+export {
+  CommandResult,
+  SessionType,
+  SessionMode,
+  SessionMetadata,
+  CommandRequest,
+  ConnectionInfo,
+} from './types.js';
+export { TIMEOUTS, BUFFER_LIMITS, SSH_CONFIG } from './constants.js';
